move tensor product
