+++ mitiq/utils.py
@@ -5,6 +5,7 @@
 
 """Utility functions."""
 from copy import deepcopy
+from itertools import product
 from typing import Any, Dict, List, Tuple
 
 import cirq
@@ -24,6 +25,7 @@ from cirq import (
     ops,
 )
 from cirq.ops.measurement_gate import MeasurementGate
+from numpy.typing import NDArray
 
 
 def _simplify_gate_exponent(gate: EigenGate) -> EigenGate:
@@ -272,7 +274,7 @@ def _safe_sqrt(
     return square_root
 
 
-def tensor_product(
+def arbitrary_tensor_product(
     *args: npt.NDArray[np.complex64],
 ) -> npt.NDArray[np.complex64]:
     """Returns the Kronecker product of the input array-like arguments.
@@ -312,3 +314,50 @@ def vector_to_matrix(
     )
     dim = _safe_sqrt(vector.size, error_str)
     return vector.reshape(dim, dim)
+
+
+PAULIS = [
+    cirq.I._unitary_(),
+    cirq.X._unitary_(),
+    cirq.Y._unitary_(),
+    cirq.Z._unitary_(),
+]
+
+
+def matrix_kronecker_product(matrices: List[NDArray[Any]]) -> NDArray[Any]:
+    """
+    Returns the Kronecker product of a list of matrices.
+    Args:
+        matrices: A list of matrices.
+    Returns:
+        The Kronecker product of the matrices in the list.
+    """
+    result = matrices[0]
+    for matrix in matrices[1:]:
+        result = np.kron(result, matrix)
+    return result
+
+
+def operator_ptm_vector_rep(opt: NDArray[Any]) -> NDArray[Any]:
+    r"""
+    Returns the PTM vector representation of an operator.
+    :math:`\mathcal{L}(\mathcal{H}_{2^n})\ni \mathtt{opt}\rightarrow
+    |\mathtt{opt}\rangle\!\rangle\in \mathcal{H}_{4^n}`.
+
+    Args:
+        opt: A square matrix representing an operator.
+    Returns:
+        A Pauli transfer matrix (PTM) representation of the operator.
+    """
+    # vector i-th entry is math:`d^{-1/2}Tr(oP_i)`
+    # where P_i is the i-th Pauli matrix
+    if not (len(opt.shape) == 2 and opt.shape[0] == opt.shape[1]):
+        raise TypeError("Input must be a square matrix")
+    num_qubits = int(np.log2(opt.shape[0]))
+    opt_vec = []
+    for pauli_combination in product(PAULIS, repeat=num_qubits):
+        kron_product = matrix_kronecker_product(pauli_combination)
+        opt_vec.append(
+            np.trace(opt @ kron_product) * np.sqrt(1 / 2**num_qubits)
+        )
+    return np.array(opt_vec)
